Move regiment class in regiment.py
