+++ W40K/Lists/Tactics_list.py
@@ -1,4 +1,4 @@
-from W40K.Class.Tactics_class import Tactic
+from W40K.Class.Tactics import Tactic
 
 ## ATTACKS
 ATK_tactics = [
